requirements.txt
Django==3.*
djangorestframework==3.11.*
markdown==3.1.*